Move the common writeHead to IndexUtils
1. move writeHead to IndexUtils, since it's common for all index writers
2. add one test case to verify IndexUtils.writeHead
3. remove some unnecessary code in Bitmap index r/w files
